nst_app/requirements_app.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.3.0
blinker==1.9.0
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
cachetools==6.1.0
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
colorama==0.4.6
efficientnet_pytorch==0.7.1
filelock @ file:///C:/b/abs_758187j281/croot/filelock_1744281404850/work
flatbuffers==25.2.10
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.14.0
huggingface-hub==0.34.4
idna @ file:///C:/b/abs_aad84bnnw5/croot/idna_1714398896795/work
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
keras==3.11.2
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.3
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
munch==4.0.0
namex==0.1.0
narwhals==2.1.2
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.1
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
pretrainedmodels==0.7.4
protobuf==5.29.5
pyarrow==21.0.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydeck==0.9.1
Pygments==2.19.2
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
referencing==0.36.2
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich==14.1.0
rpds-py==0.27.0
safetensors==0.6.2
segmentation-models-pytorch==0.3.3
six==1.17.0
smmap==5.0.2
streamlit==1.48.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
timm==0.9.2
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
watchdog==6.0.0
Werkzeug==3.1.3
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt==1.17.3
